All time-related types are changed from double to long.

+++ app/src/main/java/org/cap/model/CoreState.java
@@ -4,7 +4,7 @@ public class CoreState {
     public Task currentTask;
     public boolean isRunning;
     public int remainingRuntime;
-    public double minimumVirtualRuntime = 0;
+    public long minimumVirtualRuntime = 0L;
 
     public CoreState copy() {
         CoreState newState = new CoreState();
