+++ app/src/main/java/org/cap/model/SimulationResult.java
@@ -4,9 +4,9 @@ import java.util.HashMap;
 
 public class SimulationResult {
     public boolean schedulability;
-    public HashMap<Integer, Double> wcrtMap;
+    public HashMap<Integer, Long> wcrtMap;
 
-    public SimulationResult(boolean schedulability, HashMap<Integer, Double> wcrtMap) {
+    public SimulationResult(boolean schedulability, HashMap<Integer, Long> wcrtMap) {
         this.schedulability = schedulability;
         this.wcrtMap = wcrtMap;
     }
